Add BaseOneServerPerTest
